@@ -1,22 +1,24 @@
 # coding: utf-8
+# frozen_string_literal: true
 
 require 'nkf'
 
 RSpec.describe OAuth2::Client do
   subject do
-    described_class.new('abc', 'def', {:site => 'https://api.example.com'}.merge(options)) do |builder|
+    described_class.new('abc', 'def', {site: 'https://api.example.com'}.merge(options)) do |builder|
       builder.adapter :test do |stub|
-        stub.get('/success')             { |env| [200, {'Content-Type' => 'text/awesome'}, 'yay'] }
+        stub.get('/success')             { |_env| [200, {'Content-Type' => 'text/awesome'}, 'yay'] }
         stub.get('/reflect')             { |env| [200, {}, env[:body]] }
         stub.post('/reflect')            { |env| [200, {}, env[:body]] }
-        stub.get('/unauthorized')        { |env| [401, {'Content-Type' => 'application/json'}, MultiJson.encode(:error => error_value, :error_description => error_description_value)] }
-        stub.get('/conflict')            { |env| [409, {'Content-Type' => 'text/plain'}, 'not authorized'] }
-        stub.get('/redirect')            { |env| [302, {'Content-Type' => 'text/plain', 'location' => '/success'}, ''] }
-        stub.post('/redirect')           { |env| [303, {'Content-Type' => 'text/plain', 'location' => '/reflect'}, ''] }
-        stub.get('/error')               { |env| [500, {'Content-Type' => 'text/plain'}, 'unknown error'] }
-        stub.get('/empty_get')           { |env| [204, {}, nil] }
-        stub.get('/different_encoding')  { |env| [500, {'Content-Type' => 'application/json'}, NKF.nkf('-We', MultiJson.encode(:error => error_value, :error_description => '∞'))] }
-        stub.get('/ascii_8bit_encoding') { |env| [500, {'Content-Type' => 'application/json'}, MultiJson.encode(:error => 'invalid_request', :error_description => 'é').force_encoding('ASCII-8BIT')] }
+        stub.get('/unauthorized')        { |_env| [401, {'Content-Type' => 'application/json'}, MultiJson.encode(error: error_value, error_description: error_description_value)] }
+        stub.get('/conflict')            { |_env| [409, {'Content-Type' => 'text/plain'}, 'not authorized'] }
+        stub.get('/redirect')            { |_env| [302, {'Content-Type' => 'text/plain', 'location' => '/success'}, ''] }
+        stub.get('/redirect_no_loc')     { |_env| [302, {'Content-Type' => 'text/plain'}, ''] }
+        stub.post('/redirect')           { |_env| [303, {'Content-Type' => 'text/plain', 'location' => '/reflect'}, ''] }
+        stub.get('/error')               { |_env| [500, {'Content-Type' => 'text/plain'}, 'unknown error'] }
+        stub.get('/empty_get')           { |_env| [204, {}, nil] }
+        stub.get('/different_encoding')  { |_env| [500, {'Content-Type' => 'application/json'}, NKF.nkf('-We', MultiJson.encode(error: error_value, error_description: '∞'))] }
+        stub.get('/ascii_8bit_encoding') { |_env| [500, {'Content-Type' => 'application/json'}, MultiJson.encode(error: 'invalid_request', error_description: 'é').force_encoding('ASCII-8BIT')] }
       end
     end
   end
@@ -61,33 +63,33 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'allows true/false for raise_errors option' do
-      client = described_class.new('abc', 'def', :site => 'https://api.example.com', :raise_errors => false)
+      client = described_class.new('abc', 'def', site: 'https://api.example.com', raise_errors: false)
       expect(client.options[:raise_errors]).to be false
-      client = described_class.new('abc', 'def', :site => 'https://api.example.com', :raise_errors => true)
+      client = described_class.new('abc', 'def', site: 'https://api.example.com', raise_errors: true)
       expect(client.options[:raise_errors]).to be true
     end
 
     it 'allows override of raise_errors option' do
-      client = described_class.new('abc', 'def', :site => 'https://api.example.com', :raise_errors => true) do |builder|
+      client = described_class.new('abc', 'def', site: 'https://api.example.com', raise_errors: true) do |builder|
         builder.adapter :test do |stub|
-          stub.get('/notfound') { |env| [404, {}, nil] }
+          stub.get('/notfound') { |_env| [404, {}, nil] }
         end
       end
       expect(client.options[:raise_errors]).to be true
       expect { client.request(:get, '/notfound') }.to raise_error(OAuth2::Error)
-      response = client.request(:get, '/notfound', :raise_errors => false)
+      response = client.request(:get, '/notfound', raise_errors: false)
       expect(response.status).to eq(404)
     end
 
     it 'allows get/post for access_token_method option' do
-      client = described_class.new('abc', 'def', :site => 'https://api.example.com', :access_token_method => :get)
+      client = described_class.new('abc', 'def', site: 'https://api.example.com', access_token_method: :get)
       expect(client.options[:access_token_method]).to eq(:get)
-      client = described_class.new('abc', 'def', :site => 'https://api.example.com', :access_token_method => :post)
+      client = described_class.new('abc', 'def', site: 'https://api.example.com', access_token_method: :post)
       expect(client.options[:access_token_method]).to eq(:post)
     end
 
     it 'does not mutate the opts hash argument' do
-      opts = {:site => 'http://example.com/'}
+      opts = {site: 'http://example.com/'}
       opts2 = opts.dup
       described_class.new 'abc', 'def', opts
       expect(opts).to eq(opts2)
@@ -112,7 +114,7 @@ RSpec.describe OAuth2::Client do
 
       context 'when a URL with path is used in the site' do
         let(:options) do
-          {:site => 'https://example.com/blog'}
+          {site: 'https://example.com/blog'}
         end
 
         it 'generates an authorization URL relative to the site' do
@@ -138,7 +140,7 @@ RSpec.describe OAuth2::Client do
       end
 
       it 'does not add the redirect_uri param to the auth_code token exchange request' do
-        client = described_class.new('abc', 'def', :site => 'https://api.example.com', :auth_scheme => :request_body) do |builder|
+        client = described_class.new('abc', 'def', site: 'https://api.example.com', auth_scheme: :request_body) do |builder|
           builder.adapter :test do |stub|
             stub.post('/oauth/token', auth_code_params) do
               [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
@@ -157,7 +159,7 @@ RSpec.describe OAuth2::Client do
       end
 
       it 'adds the redirect_uri param to the auth_code token exchange request' do
-        client = described_class.new('abc', 'def', :redirect_uri => 'https://site.com/oauth/callback', :site => 'https://api.example.com', :auth_scheme => :request_body) do |builder|
+        client = described_class.new('abc', 'def', redirect_uri: 'https://site.com/oauth/callback', site: 'https://api.example.com', auth_scheme: :request_body) do |builder|
           builder.adapter :test do |stub|
             stub.post('/oauth/token', auth_code_params.merge('redirect_uri' => 'https://site.com/oauth/callback')) do
               [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
@@ -171,60 +173,60 @@ RSpec.describe OAuth2::Client do
     describe 'custom headers' do
       context 'string key headers' do
         it 'adds the custom headers to request' do
-          client = described_class.new('abc', 'def', :site => 'https://api.example.com', :auth_scheme => :request_body) do |builder|
+          client = described_class.new('abc', 'def', site: 'https://api.example.com', auth_scheme: :request_body) do |builder|
             builder.adapter :test do |stub|
               stub.post('/oauth/token') do |env|
-                expect(env.request_headers).to include({'CustomHeader' => 'CustomHeader'})
+                expect(env.request_headers).to include('CustomHeader' => 'CustomHeader')
                 [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
               end
             end
           end
-          header_params = {'headers' => { 'CustomHeader' => 'CustomHeader' }}
+          header_params = {'headers' => {'CustomHeader' => 'CustomHeader'}}
           client.auth_code.get_token('code', header_params)
         end
       end
 
       context 'symbol key headers' do
         it 'adds the custom headers to request' do
-          client = described_class.new('abc', 'def', :site => 'https://api.example.com', :auth_scheme => :request_body) do |builder|
+          client = described_class.new('abc', 'def', site: 'https://api.example.com', auth_scheme: :request_body) do |builder|
             builder.adapter :test do |stub|
               stub.post('/oauth/token') do |env|
-                expect(env.request_headers).to include({'CustomHeader' => 'CustomHeader'})
+                expect(env.request_headers).to include('CustomHeader' => 'CustomHeader')
                 [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
               end
             end
           end
-          header_params = {headers: { 'CustomHeader' => 'CustomHeader' }}
+          header_params = {headers: {'CustomHeader' => 'CustomHeader'}}
           client.auth_code.get_token('code', header_params)
         end
       end
 
       context 'string key custom headers with basic auth' do
         it 'adds the custom headers to request' do
-          client = described_class.new('abc', 'def', :site => 'https://api.example.com') do |builder|
+          client = described_class.new('abc', 'def', site: 'https://api.example.com') do |builder|
             builder.adapter :test do |stub|
               stub.post('/oauth/token') do |env|
-                expect(env.request_headers).to include({'CustomHeader' => 'CustomHeader'})
+                expect(env.request_headers).to include('CustomHeader' => 'CustomHeader')
                 [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
               end
             end
           end
-          header_params = {'headers' => { 'CustomHeader' => 'CustomHeader' }}
+          header_params = {'headers' => {'CustomHeader' => 'CustomHeader'}}
           client.auth_code.get_token('code', header_params)
         end
       end
 
       context 'symbol key custom headers with basic auth' do
         it 'adds the custom headers to request' do
-          client = described_class.new('abc', 'def', :site => 'https://api.example.com') do |builder|
+          client = described_class.new('abc', 'def', site: 'https://api.example.com') do |builder|
             builder.adapter :test do |stub|
               stub.post('/oauth/token') do |env|
-                expect(env.request_headers).to include({'CustomHeader' => 'CustomHeader'})
+                expect(env.request_headers).to include('CustomHeader' => 'CustomHeader')
                 [200, {'Content-Type' => 'application/json'}, '{"access_token":"token"}']
               end
             end
           end
-          header_params = {headers: { 'CustomHeader' => 'CustomHeader' }}
+          header_params = {headers: {'CustomHeader' => 'CustomHeader'}}
           client.auth_code.get_token('code', header_params)
         end
       end
@@ -237,6 +239,7 @@ RSpec.describe OAuth2::Client do
       before do
         stub_env('OAUTH_DEBUG' => debug_value)
       end
+
       context 'when OAUTH_DEBUG=true' do
         let(:debug_value) { 'true' }
 
@@ -251,10 +254,11 @@ RSpec.describe OAuth2::Client do
             subject.request(:get, '/success')
           end.not_to raise_error
         end
+
         it 'prints both request and response bodies to STDOUT' do
           printed = capture(:stdout) do
             subject.request(:get, '/success')
-            subject.request(:get, '/reflect', :body => 'this is magical')
+            subject.request(:get, '/reflect', body: 'this is magical')
           end
           expect(printed).to match 'request: GET https://api.example.com/success'
           expect(printed).to match 'response: Content-Type:'
@@ -263,6 +267,7 @@ RSpec.describe OAuth2::Client do
           expect(printed).to match 'response: this is magical'
         end
       end
+
       context 'when OAUTH_DEBUG=false' do
         let(:debug_value) { 'false' }
 
@@ -277,10 +282,11 @@ RSpec.describe OAuth2::Client do
             subject.request(:get, '/success')
           end.not_to raise_error
         end
+
         it 'prints nothing to STDOUT' do
           printed = capture(:stdout) do
             subject.request(:get, '/success')
-            subject.request(:get, '/reflect', :body => 'this is magical')
+            subject.request(:get, '/reflect', body: 'this is magical')
           end
           expect(printed).to eq ''
         end
@@ -300,50 +306,21 @@ RSpec.describe OAuth2::Client do
       expect(response.headers).to eq('Content-Type' => 'text/awesome')
     end
 
-    context 'when OAUTH_DEBUG=true and logger is set to log to /dev/null' do
-      around do |example|
-        begin
-          original = ENV['OAUTH_DEBUG']
-          ENV['OAUTH_DEBUG'] = 'true'
-
-          original_logger = subject.options[:logger]
-          subject.options[:logger] = Logger.new('/dev/null')
-
-          example.call
-        ensure
-          subject.options[:logger] = original_logger
-
-          if original.nil?
-            ENV.delete('OAUTH_DEBUG')
-          else
-            ENV['OAUTH_DEBUG'] = original
-          end
-        end
-      end
-
-      it 'will not log anything to standard out if logger is overridden to use /dev/null' do
-        output = capture(:stdout) do
-          subject.request(:get, '/success')
-        end
-
-        expect(output).to be_empty
-      end
-    end
-
     context 'with ENV' do
       include_context 'with stubbed env'
       context 'when OAUTH_DEBUG=true' do
         before do
           stub_env('OAUTH_DEBUG' => 'true')
         end
+
         it 'outputs to $stdout when OAUTH_DEBUG=true' do
           output = capture(:stdout) do
             subject.request(:get, '/success')
           end
           logs = [
-              'request: GET https://api.example.com/success',
-              'response: Status 200',
-              'response: Content-Type: "text/awesome"'
+            'request: GET https://api.example.com/success',
+            'response: Status 200',
+            'response: Content-Type: "text/awesome"',
           ]
           expect(output).to include(*logs)
         end
@@ -351,7 +328,7 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'posts a body' do
-      response = subject.request(:post, '/reflect', :body => 'foo=bar')
+      response = subject.request(:post, '/reflect', body: 'foo=bar')
       expect(response.body).to eq('foo=bar')
     end
 
@@ -360,12 +337,18 @@ RSpec.describe OAuth2::Client do
       expect(response.body).to eq('yay')
       expect(response.status).to eq(200)
       expect(response.headers).to eq('Content-Type' => 'text/awesome')
+      expect(response.response.env.url.to_s).to eq('https://api.example.com/success')
     end
 
     it 'redirects using GET on a 303' do
-      response = subject.request(:post, '/redirect', :body => 'foo=bar')
+      response = subject.request(:post, '/redirect', body: 'foo=bar')
       expect(response.body).to be_empty
       expect(response.status).to eq(200)
+      expect(response.response.env.url.to_s).to eq('https://api.example.com/reflect')
+    end
+
+    it 'raises an error if a redirect has no Location header' do
+      expect { subject.request(:get, '/redirect_no_loc') }.to raise_error(OAuth2::Error, 'Got 302 status code, but no Location header was present')
     end
 
     it 'obeys the :max_redirects option' do
@@ -386,6 +369,7 @@ RSpec.describe OAuth2::Client do
 
     %w[/unauthorized /conflict /error /different_encoding /ascii_8bit_encoding].each do |error_path|
       it "raises OAuth2::Error on error response to path #{error_path}" do
+        pending_for(engine: 'jruby', reason: 'https://github.com/jruby/jruby/issues/4921') if error_path == '/different_encoding'
         expect { subject.request(:get, error_path) }.to raise_error(OAuth2::Error)
       end
     end
@@ -398,22 +382,34 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'parses OAuth2 standard error response' do
-      begin
-        subject.request(:get, '/unauthorized')
-      rescue StandardError => e
-        expect(e.code).to eq(error_value)
-        expect(e.description).to eq(error_description_value)
-        expect(e.to_s).to match(/#{error_value}/)
-        expect(e.to_s).to match(/#{error_description_value}/)
+      expect { subject.request(:get, '/unauthorized') }.to raise_error do |ex|
+        expect(ex.code).to eq(error_value)
+        expect(ex.description).to eq(error_description_value)
+        expect(ex.to_s).to match(/#{error_value}/)
+        expect(ex.to_s).to match(/#{error_description_value}/)
       end
     end
 
     it 'provides the response in the Exception' do
-      begin
-        subject.request(:get, '/error')
-      rescue StandardError => e
-        expect(e.response).to be_a(OAuth2::Response)
-        expect(e.to_s).to match(/unknown error/)
+      expect { subject.request(:get, '/error') }.to raise_error do |ex|
+        expect(ex.response).not_to be_nil
+        expect(ex.to_s).to match(/unknown error/)
+      end
+    end
+
+    context 'when errors are raised by Faraday' do
+      let(:connection) { instance_double(Faraday::Connection, build_url: double) }
+
+      it 'rescues Faraday::ConnectionFailed' do
+        allow(connection).to(
+          receive(:run_request).and_raise(Faraday::ConnectionFailed.new('fail'))
+        )
+        allow(subject).to receive(:connection).and_return(connection) # rubocop:disable RSpec/SubjectStub
+
+        expect { subject.request(:get, '/redirect') }.to raise_error do |e|
+          expect(e.class).to eq(OAuth2::ConnectionError)
+          expect(e.message).to eq('fail')
+        end
       end
     end
   end
@@ -432,8 +428,8 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'authenticates with request parameters' do
-      client = stubbed_client(:auth_scheme => :request_body) do |stub|
-        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def') do |env|
+      client = stubbed_client(auth_scheme: :request_body) do |stub|
+        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def') do |_env|
           [200, {'Content-Type' => 'application/json'}, MultiJson.encode('access_token' => 'the-token')]
         end
       end
@@ -441,9 +437,10 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'authenticates with Basic auth' do
-      client = stubbed_client(:auth_scheme => :basic_auth) do |stub|
+      client = stubbed_client(auth_scheme: :basic_auth) do |stub|
         stub.post('/oauth/token') do |env|
           raise Faraday::Adapter::Test::Stubs::NotFound unless env[:request_headers]['Authorization'] == OAuth2::Authenticator.encode_basic_auth('abc', 'def')
+
           [200, {'Content-Type' => 'application/json'}, MultiJson.encode('access_token' => 'the-token')]
         end
       end
@@ -464,14 +461,18 @@ RSpec.describe OAuth2::Client do
 
     context 'when the :raise_errors flag is set to false' do
       context 'when the request body is nil' do
-        it 'returns a nil :access_token' do
-          client = stubbed_client(:raise_errors => false) do |stub|
+        subject(:get_token) { client.get_token({}) }
+
+        let(:client) do
+          stubbed_client(raise_errors: false) do |stub|
             stub.post('/oauth/token') do
               [500, {'Content-Type' => 'application/json'}, nil]
             end
           end
+        end
 
-          expect(client.get_token({})).to eq(nil)
+        it 'raises error MultiJson::ParseError' do
+          block_is_expected { get_token }.to raise_error(MultiJson::ParseError)
         end
       end
 
@@ -491,16 +492,27 @@ RSpec.describe OAuth2::Client do
     end
 
     it 'forwards given token parameters' do
-      client = stubbed_client(:auth_scheme => :request_body) do |stub|
-        stub.post('/oauth/token', 'arbitrary' => 'parameter', 'client_id' => 'abc', 'client_secret' => 'def') do |env|
+      client = stubbed_client(auth_scheme: :request_body) do |stub|
+        stub.post('/oauth/token', 'arbitrary' => 'parameter', 'client_id' => 'abc', 'client_secret' => 'def') do |_env|
           [200, {'Content-Type' => 'application/json'}, MultiJson.encode('access_token' => 'the-token')]
         end
       end
       client.get_token('arbitrary' => 'parameter')
     end
 
+    context 'when token_method is set to post_with_query_string' do
+      it 'uses the http post method and passes parameters in the query string' do
+        client = stubbed_client(token_method: :post_with_query_string) do |stub|
+          stub.post('/oauth/token?state=abc123') do |_env|
+            [200, {'Content-Type' => 'application/json'}, MultiJson.encode('access_token' => 'the-token')]
+          end
+        end
+        client.get_token('state' => 'abc123')
+      end
+    end
+
     def stubbed_client(params = {}, &stubs)
-      params = {:site => 'https://api.example.com'}.merge(params)
+      params = {site: 'https://api.example.com'}.merge(params)
       OAuth2::Client.new('abc', 'def', params) do |builder|
         builder.adapter :test, &stubs
       end
@@ -517,8 +529,8 @@ RSpec.describe OAuth2::Client do
 
   context 'with SSL options' do
     subject do
-      cli = described_class.new('abc', 'def', :site => 'https://api.example.com', :ssl => {:ca_file => 'foo.pem'})
-      cli.connection.build do |b|
+      cli = described_class.new('abc', 'def', site: 'https://api.example.com', ssl: {ca_file: 'foo.pem'})
+      cli.connection = Faraday.new(cli.site, cli.options[:connection_opts]) do |b|
         b.adapter :test
       end
       cli
@@ -531,7 +543,7 @@ RSpec.describe OAuth2::Client do
 
   context 'without a connection-configuration block' do
     subject do
-      described_class.new('abc', 'def', :site => 'https://api.example.com')
+      described_class.new('abc', 'def', site: 'https://api.example.com')
     end
 
     it 'applies default faraday middleware to the connection' do
