@@ -1,10 +1,12 @@
+# frozen_string_literal: true
+
 RSpec.describe OAuth2::Response do
-  let(:raw_response) { Faraday::Response.new(:status => status, :response_headers => headers, :body => body) }
+  subject { described_class.new(raw_response) }
+
+  let(:raw_response) { Faraday::Response.new(status: status, response_headers: headers, body: body) }
   let(:status) { 200 }
   let(:headers) { {'foo' => 'bar'} }
   let(:body) { 'foo' }
-
-  let(:subject) { described_class.new(raw_response) }
 
   describe '#initialize' do
     it 'returns the status, headers and body' do
@@ -16,13 +18,13 @@ RSpec.describe OAuth2::Response do
 
   describe '.register_parser' do
     let(:response) do
-      double('response', :headers => {'Content-Type' => 'application/foo-bar'},
-                         :status => 200,
-                         :body => 'baz')
+      double('response', headers: {'Content-Type' => 'application/foo-bar'},
+                         status: 200,
+                         body: 'baz')
     end
 
     before do
-      described_class.register_parser(:foobar, 'application/foo-bar') do |body|
+      described_class.register_parser(:foobar, ['application/foo-bar']) do |body|
         "foobar #{body}"
       end
     end
@@ -64,67 +66,140 @@ RSpec.describe OAuth2::Response do
   end
 
   describe '#parsed' do
-    it 'parses application/x-www-form-urlencoded body' do
-      headers = {'Content-Type' => 'application/x-www-form-urlencoded'}
-      body = 'foo=bar&answer=42'
-      response = double('response', :headers => headers, :body => body)
-      subject = described_class.new(response)
-      expect(subject.parsed.keys.size).to eq(2)
-      expect(subject.parsed['foo']).to eq('bar')
-      expect(subject.parsed['answer']).to eq('42')
+    subject(:parsed) do
+      headers = {'Content-Type' => content_type}
+      response = double('response', headers: headers, body: body)
+      instance = described_class.new(response)
+      instance.parsed
     end
 
-    it 'parses application/json body' do
-      headers = {'Content-Type' => 'application/json'}
-      body = MultiJson.encode(:foo => 'bar', :answer => 42)
-      response = double('response', :headers => headers, :body => body)
-      subject = described_class.new(response)
-      expect(subject.parsed.keys.size).to eq(2)
-      expect(subject.parsed['foo']).to eq('bar')
-      expect(subject.parsed['answer']).to eq(42)
+    shared_examples_for 'parsing JSON-like' do
+      it 'has num keys' do
+        expect(parsed.keys.size).to eq(6)
+      end
+
+      it 'parses string' do
+        expect(parsed['foo']).to eq('bar')
+        expect(parsed.key('bar')).to eq('foo')
+      end
+
+      it 'parses non-zero number' do
+        expect(parsed['answer']).to eq(42)
+        expect(parsed.key(42)).to eq('answer')
+      end
+
+      it 'parses nil as NilClass' do
+        expect(parsed['krill']).to be_nil
+        expect(parsed.key(nil)).to eq('krill')
+      end
+
+      it 'parses zero as number' do
+        expect(parsed['zero']).to eq(0)
+        expect(parsed.key(0)).to eq('zero')
+      end
+
+      it 'parses false as FalseClass' do
+        expect(parsed['malign']).to be(false)
+        expect(parsed.key(false)).to eq('malign')
+      end
+
+      it 'parses false as TrueClass' do
+        expect(parsed['shine']).to be(true)
+        expect(parsed.key(true)).to eq('shine')
+      end
     end
 
-    it 'parses alternative application/json extension bodies' do
-      headers = {'Content-Type' => 'application/hal+json'}
-      body = MultiJson.encode(:foo => 'bar', :answer => 42)
-      response = double('response', :headers => headers, :body => body)
-      subject = described_class.new(response)
-      expect(subject.parsed.keys.size).to eq(2)
-      expect(subject.parsed['foo']).to eq('bar')
-      expect(subject.parsed['answer']).to eq(42)
+    context 'when application/json' do
+      let(:content_type) { 'application/json' }
+      let(:body) { MultiJson.encode(foo: 'bar', answer: 42, krill: nil, zero: 0, malign: false, shine: true) }
+
+      it_behaves_like 'parsing JSON-like'
+    end
+
+    context 'when application/Json' do
+      let(:content_type) { 'application/Json' }
+      let(:body) { MultiJson.encode(foo: 'bar', answer: 42, krill: nil, zero: 0, malign: false, shine: true) }
+
+      it_behaves_like 'parsing JSON-like'
+    end
+
+    context 'when application/hal+json' do
+      let(:content_type) { 'application/hal+json' }
+      let(:body) { MultiJson.encode(foo: 'bar', answer: 42, krill: nil, zero: 0, malign: false, shine: true) }
+
+      it_behaves_like 'parsing JSON-like'
+    end
+
+    context 'when application/x-www-form-urlencoded' do
+      let(:content_type) { 'application/x-www-form-urlencoded' }
+      let(:body) { 'foo=bar&answer=42&krill=&zero=0&malign=false&shine=true' }
+
+      it 'has num keys' do
+        expect(parsed.keys.size).to eq(6)
+      end
+
+      it 'parses string' do
+        expect(parsed['foo']).to eq('bar')
+        expect(parsed.key('bar')).to eq('foo')
+      end
+
+      it 'parses non-zero number as string' do
+        expect(parsed['answer']).to eq('42')
+        expect(parsed.key('42')).to eq('answer')
+      end
+
+      it 'parses nil as empty string' do
+        expect(parsed['krill']).to eq('')
+        expect(parsed.key('')).to eq('krill')
+      end
+
+      it 'parses zero as string' do
+        expect(parsed['zero']).to eq('0')
+        expect(parsed.key('0')).to eq('zero')
+      end
+
+      it 'parses false as string' do
+        expect(parsed['malign']).to eq('false')
+        expect(parsed.key('false')).to eq('malign')
+      end
+
+      it 'parses true as string' do
+        expect(parsed['shine']).to eq('true')
+        expect(parsed.key('true')).to eq('shine')
+      end
     end
 
     it 'parses application/vnd.collection+json body' do
       headers = {'Content-Type' => 'application/vnd.collection+json'}
-      body = MultiJson.encode(:collection => {})
-      response = double('response', :headers => headers, :body => body)
+      body = MultiJson.encode(collection: {})
+      response = double('response', headers: headers, body: body)
       subject = described_class.new(response)
       expect(subject.parsed.keys.size).to eq(1)
     end
 
     it 'parses application/vnd.api+json body' do
       headers = {'Content-Type' => 'application/vnd.api+json'}
-      body = MultiJson.encode(:collection => {})
-      response = double('response', :headers => headers, :body => body)
+      body = MultiJson.encode(collection: {})
+      response = double('response', headers: headers, body: body)
       subject = described_class.new(response)
       expect(subject.parsed.keys.size).to eq(1)
     end
 
-    it "parses application/Json body" do
-      headers = {'Content-Type' => 'application/Json'}
-      body = MultiJson.encode(:foo => 'bar', :answer => 42)
-      response = double('response', :headers => headers, :body => body)
-      subject = Response.new(response)
+    it 'parses application/problem+json body' do
+      headers = {'Content-Type' => 'application/problem+json'}
+      body = MultiJson.encode(type: 'https://tools.ietf.org/html/rfc7231#section-6.5.4', title: 'Not Found')
+      response = double('response', headers: headers, body: body)
+      subject = described_class.new(response)
       expect(subject.parsed.keys.size).to eq(2)
-      expect(subject.parsed['foo']).to eq('bar')
-      expect(subject.parsed['answer']).to eq(42)
+      expect(subject.parsed['type']).to eq('https://tools.ietf.org/html/rfc7231#section-6.5.4')
+      expect(subject.parsed['title']).to eq('Not Found')
     end
 
     it "doesn't try to parse other content-types" do
       headers = {'Content-Type' => 'text/html'}
       body = '<!DOCTYPE html><html><head></head><body></body></html>'
 
-      response = double('response', :headers => headers, :body => body)
+      response = double('response', headers: headers, body: body)
 
       expect(MultiJson).not_to receive(:decode)
       expect(MultiJson).not_to receive(:load)
@@ -134,14 +209,14 @@ RSpec.describe OAuth2::Response do
       expect(subject.parsed).to be_nil
     end
 
-    it "should snakecase json keys when parsing" do
+    it 'snakecases json keys when parsing' do
       headers = {'Content-Type' => 'application/json'}
-      body = MultiJson.encode("accessToken" => 'bar', "MiGever" => "Ani")
-      response = double('response', :headers => headers, :body => body)
-      subject = Response.new(response)
+      body = MultiJson.encode('accessToken' => 'bar', 'MiGever' => 'Ani')
+      response = double('response', headers: headers, body: body)
+      subject = described_class.new(response)
       expect(subject.parsed.keys.size).to eq(2)
       expect(subject.parsed['access_token']).to eq('bar')
-      expect(subject.parsed['mi_gever']).to eq("Ani")
+      expect(subject.parsed['mi_gever']).to eq('Ani')
     end
 
     it 'supports registered parsers with arity == 0; passing nothing' do
@@ -151,9 +226,9 @@ RSpec.describe OAuth2::Response do
 
       headers   = {'Content-Type' => 'text/html'}
       body      = '<!DOCTYPE html><html><head></head><body></body></html>'
-      response  = double('response', :headers => headers, :body => body)
+      response  = double('response', headers: headers, body: body)
 
-      subject = described_class.new(response, :parse => :arity_0)
+      subject = described_class.new(response, parse: :arity_0)
 
       expect(subject.parsed).to eq('a-ok')
     end
@@ -161,7 +236,7 @@ RSpec.describe OAuth2::Response do
     it 'supports registered parsers with arity == 2; passing body and response' do
       headers   = {'Content-Type' => 'text/html'}
       body      = '<!DOCTYPE html><html><head></head><body></body></html>'
-      response  = double('response', :headers => headers, :body => body)
+      response  = double('response', headers: headers, body: body)
 
       described_class.register_parser(:arity_2, []) do |passed_body, passed_response|
         expect(passed_body).to eq(body)
@@ -170,7 +245,7 @@ RSpec.describe OAuth2::Response do
         'a-ok'
       end
 
-      subject = described_class.new(response, :parse => :arity_2)
+      subject = described_class.new(response, parse: :arity_2)
 
       expect(subject.parsed).to eq('a-ok')
     end
@@ -178,7 +253,7 @@ RSpec.describe OAuth2::Response do
     it 'supports registered parsers with arity > 2; passing body and response' do
       headers   = {'Content-Type' => 'text/html'}
       body      = '<!DOCTYPE html><html><head></head><body></body></html>'
-      response  = double('response', :headers => headers, :body => body)
+      response  = double('response', headers: headers, body: body)
 
       described_class.register_parser(:arity_3, []) do |passed_body, passed_response, *args|
         expect(passed_body).to eq(body)
@@ -188,7 +263,7 @@ RSpec.describe OAuth2::Response do
         'a-ok'
       end
 
-      subject = described_class.new(response, :parse => :arity_3)
+      subject = described_class.new(response, parse: :arity_3)
 
       expect(subject.parsed).to eq('a-ok')
     end
@@ -196,16 +271,16 @@ RSpec.describe OAuth2::Response do
     it 'supports directly passed parsers' do
       headers   = {'Content-Type' => 'text/html'}
       body      = '<!DOCTYPE html><html><head></head><body></body></html>'
-      response  = double('response', :headers => headers, :body => body)
+      response  = double('response', headers: headers, body: body)
 
-      subject = described_class.new(response, :parse => lambda { 'a-ok' })
+      subject = described_class.new(response, parse: -> { 'a-ok' })
 
       expect(subject.parsed).to eq('a-ok')
     end
   end
 
   context 'with xml parser registration' do
-    it 'tries to load multi_xml and use it' do
+    it 'tries to load multi_xml.rb and use it' do
       expect(described_class.send(:class_variable_get, :@@parsers)[:xml]).not_to be_nil
     end
 
@@ -213,7 +288,7 @@ RSpec.describe OAuth2::Response do
       headers = {'Content-Type' => 'text/xml'}
       body = '<?xml version="1.0" standalone="yes" ?><foo><bar>baz</bar></foo>'
 
-      response = double('response', :headers => headers, :body => body)
+      response = double('response', headers: headers, body: body)
       expect(described_class.new(response).parsed).to eq('foo' => {'bar' => 'baz'})
     end
 
@@ -221,7 +296,7 @@ RSpec.describe OAuth2::Response do
       headers = {'Content-Type' => 'application/xml'}
       body = '<?xml version="1.0" standalone="yes" ?><foo><bar>baz</bar></foo>'
 
-      response = double('response', :headers => headers, :body => body)
+      response = double('response', headers: headers, body: body)
       expect(described_class.new(response).parsed).to eq('foo' => {'bar' => 'baz'})
     end
   end
