@@ -1,4 +1,5 @@
 # encoding: utf-8
+# frozen_string_literal: true
 
 RSpec.describe OAuth2::Strategy::AuthCode do
   subject { client.auth_code }
@@ -6,14 +7,14 @@ RSpec.describe OAuth2::Strategy::AuthCode do
   let(:code) { 'sushi' }
   let(:kvform_token) { 'expires_in=600&access_token=salmon&refresh_token=trout&extra_param=steve' }
   let(:facebook_token) { kvform_token.gsub('_in', '') }
-  let(:json_token) { MultiJson.encode(:expires_in => 600, :access_token => 'salmon', :refresh_token => 'trout', :extra_param => 'steve') }
+  let(:json_token) { MultiJson.encode(expires_in: 600, access_token: 'salmon', refresh_token: 'trout', extra_param: 'steve') }
   let(:redirect_uri) { 'http://example.com/redirect_uri' }
   let(:microsoft_token) { 'id_token=jwt' }
 
   let(:client) do
-    OAuth2::Client.new('abc', 'def', :site => 'http://api.example.com') do |builder|
+    OAuth2::Client.new('abc', 'def', site: 'http://api.example.com') do |builder|
       builder.adapter :test do |stub|
-        stub.get("/oauth/token?client_id=abc&code=#{code}&grant_type=authorization_code") do |env|
+        stub.get("/oauth/token?client_id=abc&code=#{code}&grant_type=authorization_code") do |_env|
           case @mode
           when 'formencoded'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, kvform_token]
@@ -23,9 +24,10 @@ RSpec.describe OAuth2::Strategy::AuthCode do
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, facebook_token]
           when 'from_microsoft'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, microsoft_token]
+          else raise ArgumentError, "Bad @mode: #{@mode}"
           end
         end
-        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'code' => 'sushi', 'grant_type' => 'authorization_code') do |env|
+        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'code' => 'sushi', 'grant_type' => 'authorization_code') do |_env|
           case @mode
           when 'formencoded'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, kvform_token]
@@ -33,9 +35,10 @@ RSpec.describe OAuth2::Strategy::AuthCode do
             [200, {'Content-Type' => 'application/json'}, json_token]
           when 'from_facebook'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, facebook_token]
+          else raise ArgumentError, "Bad @mode: #{@mode}"
           end
         end
-        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'code' => 'sushi', 'grant_type' => 'authorization_code', 'redirect_uri' => redirect_uri) do |env|
+        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'code' => 'sushi', 'grant_type' => 'authorization_code', 'redirect_uri' => redirect_uri) do |_env|
           [200, {'Content-Type' => 'application/json'}, json_token]
         end
       end
@@ -56,7 +59,7 @@ RSpec.describe OAuth2::Strategy::AuthCode do
     end
 
     it 'raises an error if the client_secret is passed in' do
-      expect { subject.authorize_url(:client_secret => 'def') }.to raise_error(ArgumentError)
+      expect { subject.authorize_url(client_secret: 'def') }.to raise_error(ArgumentError)
     end
 
     it 'raises an error if the client_secret is passed in with string keys' do
@@ -65,7 +68,7 @@ RSpec.describe OAuth2::Strategy::AuthCode do
 
     it 'includes passed in options' do
       cb = 'http://myserver.local/oauth/callback'
-      expect(subject.authorize_url(:redirect_uri => cb)).to include("redirect_uri=#{CGI.escape(cb)}")
+      expect(subject.authorize_url(redirect_uri: cb)).to include("redirect_uri=#{CGI.escape(cb)}")
     end
   end
 
@@ -78,12 +81,12 @@ RSpec.describe OAuth2::Strategy::AuthCode do
     end
 
     it 'includes redirect_uri once in the request parameters' do
-      expect { subject.get_token(code, :redirect_uri => redirect_uri) }.not_to raise_error
+      expect { subject.get_token(code, redirect_uri: redirect_uri) }.not_to raise_error
     end
   end
 
   describe '#get_token (handling utf-8 data)' do
-    let(:json_token) { MultiJson.encode(:expires_in => 600, :access_token => 'salmon', :refresh_token => 'trout', :extra_param => 'André') }
+    let(:json_token) { MultiJson.encode(expires_in: 600, access_token: 'salmon', refresh_token: 'trout', extra_param: 'André') }
 
     before do
       @mode = 'json'
@@ -113,7 +116,7 @@ RSpec.describe OAuth2::Strategy::AuthCode do
   end
 
   %w[json formencoded from_facebook].each do |mode|
-    [:get, :post].each do |verb|
+    %i[get post].each do |verb|
       describe "#get_token (#{mode}, access_token_method=#{verb}" do
         before do
           @mode = mode
