@@ -1,3 +1,5 @@
+# frozen_string_literal: true
+
 RSpec.describe OAuth2::Strategy::ClientCredentials do
   subject { client.client_credentials }
 
@@ -5,25 +7,27 @@ RSpec.describe OAuth2::Strategy::ClientCredentials do
   let(:json_token) { '{"expires_in":600,"access_token":"salmon","refresh_token":"trout"}' }
 
   let(:client) do
-    OAuth2::Client.new('abc', 'def', :site => 'http://api.example.com') do |builder|
+    OAuth2::Client.new('abc', 'def', site: 'http://api.example.com') do |builder|
       builder.adapter :test do |stub|
         stub.post('/oauth/token', 'grant_type' => 'client_credentials') do |env|
           client_id, client_secret = Base64.decode64(env[:request_headers]['Authorization'].split(' ', 2)[1]).split(':', 2)
-          client_id == 'abc' && client_secret == 'def' || raise(Faraday::Adapter::Test::Stubs::NotFound)
+          (client_id == 'abc' && client_secret == 'def') || raise(Faraday::Adapter::Test::Stubs::NotFound)
           @last_headers = env[:request_headers]
           case @mode
           when 'formencoded'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, kvform_token]
           when 'json'
             [200, {'Content-Type' => 'application/json'}, json_token]
+          else raise ArgumentError, "Bad @mode: #{@mode}"
           end
         end
-        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'grant_type' => 'client_credentials') do |env|
+        stub.post('/oauth/token', 'client_id' => 'abc', 'client_secret' => 'def', 'grant_type' => 'client_credentials') do |_env|
           case @mode
           when 'formencoded'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, kvform_token]
           when 'json'
             [200, {'Content-Type' => 'application/json'}, json_token]
+          else raise ArgumentError, "Bad @mode: #{@mode}"
           end
         end
       end
@@ -37,7 +41,7 @@ RSpec.describe OAuth2::Strategy::ClientCredentials do
   end
 
   %w[json formencoded].each do |mode|
-    [:basic_auth, :request_body].each do |auth_scheme|
+    %i[basic_auth request_body].each do |auth_scheme|
       describe "#get_token (#{mode}) (#{auth_scheme})" do
         before do
           @mode = mode
@@ -71,7 +75,7 @@ RSpec.describe OAuth2::Strategy::ClientCredentials do
   describe '#get_token (with extra header parameters)' do
     before do
       @mode = 'json'
-      @access = subject.get_token(:headers => {'X-Extra-Header' => 'wow'})
+      @access = subject.get_token(headers: {'X-Extra-Header' => 'wow'})
     end
 
     it 'sends the header correctly.' do
