@@ -1,16 +1,20 @@
+# frozen_string_literal: true
+
 RSpec.describe OAuth2::Strategy::Password do
   subject { client.password }
 
   let(:client) do
-    cli = OAuth2::Client.new('abc', 'def', :site => 'http://api.example.com')
-    cli.connection.build do |b|
+    cli = OAuth2::Client.new('abc', 'def', site: 'http://api.example.com')
+    cli.connection = Faraday.new(cli.site, cli.options[:connection_opts]) do |b|
+      b.request :url_encoded
       b.adapter :test do |stub|
-        stub.post('/oauth/token') do |env|
+        stub.post('/oauth/token') do |_env|
           case @mode
           when 'formencoded'
             [200, {'Content-Type' => 'application/x-www-form-urlencoded'}, 'expires_in=600&access_token=salmon&refresh_token=trout']
           when 'json'
             [200, {'Content-Type' => 'application/json'}, '{"expires_in":600,"access_token":"salmon","refresh_token":"trout"}']
+          else raise ArgumentError, "Bad @mode: #{@mode}"
           end
         end
       end
