@@ -1,9 +1,11 @@
-RSpec.describe SnakyHash do
+# frozen_string_literal: true
+
+RSpec.describe OAuth2::SnakyHash do
   subject { described_class.new }
 
   describe '.build' do
-    context 'build from hash' do
-      subject { described_class.build({ 'AccessToken' => '1' }) }
+    context 'when build from hash' do
+      subject { described_class.new('AccessToken' => '1') }
 
       it 'create correct snake hash' do
         expect(subject).to be_a(described_class)
@@ -12,12 +14,12 @@ RSpec.describe SnakyHash do
       end
     end
 
-    context 'build from snake_hash' do
+    context 'when build from snake_hash' do
       subject do
         h = described_class.new
         h['AccessToken'] = '1'
 
-        described_class.build(h)
+        described_class.new(h)
       end
 
       it 'create correct snake hash' do
@@ -36,14 +38,14 @@ RSpec.describe SnakyHash do
       expect(subject['access_token']).to eq('1')
     end
 
-    it 'returns assigned value with snake key only' do
+    it 'returns assigned value with snake key' do
       subject['access_token'] = '1'
 
-      expect(subject['AccessToken']).to eq(nil)
+      expect(subject['AccessToken']).to eq('1')
       expect(subject['access_token']).to eq('1')
     end
 
-    it 'overwrite snake key' do
+    it 'overwrite by alternate key' do
       subject['AccessToken'] = '1'
 
       expect(subject['AccessToken']).to eq('1')
@@ -51,81 +53,82 @@ RSpec.describe SnakyHash do
 
       subject['access_token'] = '2'
 
-      expect(subject['AccessToken']).to eq('1')
+      expect(subject['AccessToken']).to eq('2')
       expect(subject['access_token']).to eq('2')
     end
   end
 
   describe '#fetch' do
-    context 'Camel case key' do
-      subject { described_class.build('AccessToken' => '1') }
+    context 'when Camel case key' do
+      subject { described_class.new('AccessToken' => '1') }
 
       it 'return correct token' do
         expect(subject.fetch('access_token')).to eq('1')
       end
     end
 
-    context 'Camel case key with dowcased first letter' do
-      subject { described_class.build('accessToken' => '1') }
+    context 'when Camel case key with down-cased first letter' do
+      subject { described_class.new('accessToken' => '1') }
 
       it 'return correct token' do
         expect(subject.fetch('access_token')).to eq('1')
       end
     end
 
-    context 'snake case key' do
-      subject { described_class.build('access_token' => '1') }
+    context 'when snake case key' do
+      subject { described_class.new('access_token' => '1') }
 
       it 'return correct token' do
         expect(subject.fetch('access_token')).to eq('1')
       end
     end
 
-    context 'missing any key' do
+    context 'when missing any key' do
       subject { described_class.new }
 
       it 'raise KeyError with key' do
-        expect {
+        pending_for(engine: 'jruby', versions: '3.1.0', reason: 'https://github.com/jruby/jruby/issues/7112')
+        expect do
           subject.fetch('access_token')
-        }.to raise_error(KeyError, /access_token/)
+        end.to raise_error(KeyError, /access_token/)
       end
 
       it 'return default value' do
-        expect(subject.fetch('access_token') {'default'}).to eq('default')
+        expect(subject.fetch('access_token', 'default')).to eq('default')
       end
     end
   end
 
   describe '#key?' do
-   context 'Camel case key' do
-      subject { described_class.build('AccessToken' => '1') }
+    context 'when Camel case key' do
+      subject { described_class.new('AccessToken' => '1') }
 
       it 'return true' do
-        expect(subject.key?('access_token')).to eq(true)
+        expect(subject.key?('access_token')).to be(true)
       end
     end
 
-    context 'Camel case key with dowcased first letter' do
-      subject { described_class.build('accessToken' => '1') }
+    context 'when Camel case key with down-cased first letter' do
+      subject { described_class.new('accessToken' => '1') }
 
       it 'return true' do
-        expect(subject.key?('access_token')).to eq(true)
+        expect(subject.key?('access_token')).to be(true)
       end
     end
 
-    context 'snake case key' do
-      subject { described_class.build('access_token' => '1') }
+    context 'when snake case key' do
+      subject { described_class.new('access_token' => '1') }
 
       it 'return true' do
-        expect(subject.key?('access_token')).to eq(true)
+        expect(subject.key?('access_token')).to be(true)
       end
     end
 
-    context 'missing any key' do
+    context 'when missing any key' do
       subject { described_class.new }
 
       it 'return false' do
-        expect(subject.key?('access_token')).to eq(false)
+        expect(subject.key?('access_token')).to be(false)
       end
     end
   end
